Fix casting error

+++ Engine/engine/utils/RNG.h
@@ -35,9 +35,9 @@ private:
 		seed32 += 0xe120fc15;
 		uint64_t tmp;
 		tmp = (uint64_t)seed32 * 0x4a39b70d;
-		uint32_t m1 = (tmp >> 32) ^ tmp;
+		uint32_t m1 = (uint32_t)((tmp >> 32) ^ tmp);
 		tmp = (uint64_t)m1 * 0x12fad5c9;
-		uint32_t m2 = (tmp >> 32) ^ tmp;
+		uint32_t m2 = (uint32_t)((tmp >> 32) ^ tmp);
 		return m2;
 	}
 	/*
